requirements.txt
avro-python3==1.9.0
certifi==2019.3.9
chardet==3.0.4
confluent-kafka==1.0.0
fastavro==0.21.23
idna==2.8
psycopg2==2.8.2
python-dotenv==0.10.2
requests==2.22.0
urllib3==1.25.3
